MLA — compiled

+++ compiled/events/messageCreate.js
@@ -41,7 +41,7 @@ export default (client) => {
                 return;
             const dataToIncrement = args[1];
             const incrementStr = args[2];
-            const incrementValue = parseInt(incrementStr);
+            const incrementValue = parseFloat(incrementStr);
             if (invalidNumber(incrementValue))
                 return;
             await updateProfile(mention.id, {
